pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oci-metrics-mcp-server"
version = "1.0.0"
description = "MCP server for Oracle Cloud Infrastructure compute agent metrics"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "oci>=2.100.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
oci-metrics-server = "mcp_server.server:main"
